requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flask-cors==6.0.0
gunicorn==23.0.0
python-dotenv==1.1.0
requests==2.32.3
dateparser==1.2.2
SQLAlchemy==2.0.41